fix: extracts bytesize and clock
